Move to existing utils file
