+++ src/internal/commands/builtin/tidy.rs
@@ -26,6 +26,7 @@ use crate::internal::config::CommandSyntax;
 use crate::internal::config::ConfigSource;
 use crate::internal::config::SyntaxOptArg;
 use crate::internal::config::SyntaxOptArgType;
+use crate::internal::config::up::utils::directory::safe_rename;
 use crate::internal::env::shell_is_interactive;
 use crate::internal::git::format_path_with_template;
 use crate::internal::git::package_path_from_handle;
@@ -34,7 +35,6 @@ use crate::internal::git::path_entry_config;
 use crate::internal::git::safe_git_url_parse;
 use crate::internal::git_env;
 use crate::internal::user_interface::StringColor;
-use crate::internal::utils::safe_rename;
 use crate::internal::ConfigLoader;
 use crate::internal::ORG_LOADER;
 use crate::omni_error;
